+++ client/js/modules/explorer.js
@@ -1,7 +1,0 @@
-export const processExplorer = data => {
-    if (!data) return
-
-    const {blockchainLength} = data
-
-    $("#explorer-height").html(blockchainLength)
-}